Fixed grid position of entities, added editor_marker
Positioning of generic entities on the hex grid is now fixed. The code is generally restructured to separate temp debug methods from final ones. Zooming on touchpad was fixed. First attempt for logic to detect hi-dpi display and scale UI accordingly as well as by zoom level on all devices. Added editor_marker that serves to mark hex tiles to be special on some way (currently it allows to mark hexes as village, giving a name that is rendered on to the map). All nodes that represent entities on the game map should inherit the entity-script now. Units evaluate "direction" attribute. Added 'globals' as autoloaded global script to access common methods e.g. for standardized Color()-getters.