Unit class enhanced, flood fill visualized with tile origins, simple GUI
Added list of all units, each with grid local position. Enhanced flood fill to stop when target tile is found. Enhanced visualization of flood fill: Rendering an arrow onto every hex tile, pointing to its origin tile. Added some methods to check if click was on tile or unit, this allows to distinguish what is clicked by the player. Added simple GUI class that currently shows a popup if a hex tile is clicked and fills it with raw info about the corresponding tile object, refine this later. Every entity if type unit will now determine and save its grid local location on start of the game. The map (aka 'game') node is now handling the update of unit entities. This may actually be rolled back if no call-dependencies arise, where unit must call a method from game which is depending on members that are not yet initialized because ready() is for some reason called from down-to-top in the Godot scene tree. Ignored everything in 'logs' folder.